Make searchbar_js.blade into plain javascript file
